Upgrade V8 7.9.218

+++ v8/src/heap/heap.cc
@@ -4114,8 +4114,9 @@ void CollectSlots(MemoryChunk* chunk, Address start, Address end,
         }
         return KEEP_SLOT;
       },
-      SlotSet::PREFREE_EMPTY_BUCKETS);
+      SlotSet::FREE_EMPTY_BUCKETS);
   if (direction == OLD_TO_NEW) {
+    CHECK(chunk->SweepingDone());
     RememberedSetSweeping::Iterate(
         chunk,
         [start, end, untyped](MaybeObjectSlot slot) {
@@ -4124,7 +4125,7 @@ void CollectSlots(MemoryChunk* chunk, Address start, Address end,
           }
           return KEEP_SLOT;
         },
-        SlotSet::PREFREE_EMPTY_BUCKETS);
+        SlotSet::FREE_EMPTY_BUCKETS);
   }
   RememberedSet<direction>::IterateTyped(
       chunk, [=](SlotType type, Address slot) {
@@ -5560,9 +5561,10 @@ void Heap::ClearRecordedSlot(HeapObject object, ObjectSlot slot) {
   if (!page->InYoungGeneration()) {
     DCHECK_EQ(page->owner_identity(), OLD_SPACE);
 
-    store_buffer()->MoveAllEntriesToRememberedSet();
-    RememberedSet<OLD_TO_NEW>::Remove(page, slot.address());
-    RememberedSetSweeping::Remove(page, slot.address());
+    if (!page->SweepingDone()) {
+      store_buffer()->MoveAllEntriesToRememberedSet();
+      RememberedSet<OLD_TO_NEW>::Remove(page, slot.address());
+    }
   }
 #endif
 }
@@ -5591,11 +5593,11 @@ void Heap::ClearRecordedSlotRange(Address start, Address end) {
   if (!page->InYoungGeneration()) {
     DCHECK_EQ(page->owner_identity(), OLD_SPACE);
 
-    store_buffer()->MoveAllEntriesToRememberedSet();
-    RememberedSet<OLD_TO_NEW>::RemoveRange(page, start, end,
-                                           SlotSet::KEEP_EMPTY_BUCKETS);
-    RememberedSetSweeping::RemoveRange(page, start, end,
-                                       SlotSet::KEEP_EMPTY_BUCKETS);
+    if (!page->SweepingDone()) {
+      store_buffer()->MoveAllEntriesToRememberedSet();
+      RememberedSet<OLD_TO_NEW>::RemoveRange(page, start, end,
+                                             SlotSet::KEEP_EMPTY_BUCKETS);
+    }
   }
 #endif
 }
